EVALUACION 1 SOLUCIONES IA/requirements.txt
streamlit>=1.25.0
pandas>=2.1.0
numpy>=1.26.0
plotly>=5.18.0
scikit-learn>=1.3.0
openai>=1.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
pyyaml>=6.0